+++ NEWS.md
@@ -84,7 +84,7 @@
 
 # simstudy 0.1.12
 
-* Fixed genCorFlex and addMultiFac to accomodate bug fixes with package data.table
+* Fixed genCorFlex and addMultiFac to accommodate bug fixes with package data.table
 
 # simstudy 0.1.13